- Use cases - Use itunes to retrieve episodes

+++ data/src/main/java/com/caldeirasoft/castly/data/dto/itunes/LookupItemDto.kt
@@ -63,7 +63,7 @@ class LookupItemDto {
 
     class GenreResult {
         @Json(name = "genreId")
-        var genreId:Long = 0L
+        var genreId:Int = 0
 
         @Json(name = "name")
         var name:String = ""
